requirements.txt
fastapi==0.110.0
uvicorn==0.28.1
google-cloud-dialogflow==2.30.0
pymongo==4.6.2
python-dotenv==0.19.2